Use time.Timer for the election timer
Switch from using gocron for the election timeout timer, to time.Timer.
This seems to fix the issue causing the timer to stop once a leader is
found.

However, the fix requires we track when last the leader last sent a
heartbeat in all the followers. This is because the timers created using
time.AfterFunc seem to fire even when the timer is stopped.

Fixes #2
Related to #3

Signed-off-by: Jason Rogena <jason@rogena.me>